more changes

+++ lab 1/lab.c
@@ -1,4 +1,9 @@
+//lab.c (Fixed Version)
+// Ahmed Hussein 2025
 #include "lab_lists.h"
+#include <stdio.h>     
+#include <stdlib.h>     
+#include <string.h>    
 
 int main() {
     struct node* directory = NULL;
@@ -13,37 +18,35 @@ int main() {
         printf("5. Delete all records\n");
         printf("6. Exit\n");
         printf("Enter choice: ");
-        scanf("%d", &choice);
-        printf("\n\n");
+
+        //FIX:Input validation
+        if (scanf("%d", &choice) != 1) {
+            printf("Invalid input. Please enter a number.\n");
+            while (getchar() != '\n'); 
+            continue;
+        }
 
         switch (choice) {
-            case 1:
-                display(directory);
-                break;
-            case 2:
-                insert_record(&directory);
-                break;
-            case 3:
-                delete_record(&directory);
-                break;
+            case 1: display(directory); break;
+            case 2: insert_record(&directory); break;
+            case 3: delete_record(&directory); break;
             case 4:
                 printf("Search by: 1 - Name, 2 - Phone Number: ");
                 int opt;
-                scanf("%d", &opt);
+                if (scanf("%d", &opt) != 1 || (opt != 1 && opt != 2)) {
+                    printf("Invalid search option.\n");
+                    break;
+                }
                 query_directory(directory, opt);
                 break;
-            case 5:
-                delete_directory(&directory);
-                break;
+            case 5: delete_directory(&directory); break;
             case 6:
                 delete_directory(&directory);
                 printf("Exiting...\n");
                 return 0;
             default:
-                printf("Invalid choice. Try again later nerd.\n");
+                printf("Invalid choice. Please try again.\n");
         }
     }
 }
-
-
 // 2025 Ahmed Hussein
